requirements.txt
vanna
pandas
python-dotenv
streamlit
sentence-transformers
numpy
requests
matplotlib 